user/urls.py
from rest_framework.urlpatterns import format_suffix_patterns
from django.urls import path
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView, TokenVerifyView
from .views import RegisterView, CallHistoryView
urlpatterns = format_suffix_patterns([
    path('register/', RegisterView.as_view(), name='register'),
    path('login/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('verify/', TokenVerifyView.as_view(), name='token_verify'),
    path('call-history/', CallHistoryView.as_view(), name='call_history'),
])
