requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
audioop-lts==0.2.1
audioread==3.0.1
bidict==0.23.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
datasets==3.3.0
decorator==5.1.1
dill==0.3.8
Django==5.1.6
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
eventlet==0.39.0
ffmpeg-python==0.2.0
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
future==1.0.0
greenlet==3.1.1
h11==0.14.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
numba==0.61.0
numpy==2.2.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.1
pyarrow==19.0.0
PyAudio==0.2.14
pycparser==2.22
pydub==0.25.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentencepiece==0.2.0
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.48.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
